contributor_client/requirements.txt
lib-platform
python-socketio
regex
uuid
py-cpuinfo
GPUtil
pyyaml
docker